+++ centreon/src/Core/Migration/Infrastructure/API/FindMigrations/FindMigrationsRoute.yaml
@@ -1,0 +1,5 @@
+FindMigrations:
+  methods: GET
+  path: /administration/migrations
+  controller: 'Core\Migration\Infrastructure\API\FindMigrations\FindMigrationsController'
+  condition: "request.attributes.get('version') >= 24.04"
