--- centreon/src/Core/Platform/Migration/Infrastructure/API/FindMigrations/FindMigrationsRoute.yaml
@@ -1,5 +1,0 @@
-FindMigrations:
-  methods: GET
-  path: /administration/platform/migrations
-  controller: 'Core\Platform\Migration\Infrastructure\API\FindMigrations\FindMigrationsController'
-  condition: "request.attributes.get('version') >= 24.04"
